requirements.txt
black==22.1.0
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-sqlalchemy==0.27.0
mypy-extensions==0.4.3
passlib==1.7.4
pathspec==0.9.0
platformdirs==2.4.1
PyJWT==2.3.0
python-dotenv==0.19.2
six==1.16.0
SQLAlchemy==1.4.31
tomli==2.0.0
typing_extensions==4.0.1
Werkzeug==2.0.2
